legal-rag/requirements.txt
# Core dependencies
# Python 3.9+ required (not installable via pip)

# PDF Processing
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.4

# LangChain & LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-google-genai>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
